pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "model"}
packages = []

[project]
name = "esm2ae"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.7.0",
    "biopython>=1.85",
    "datasets>=3.6.0",
    "deepspeed>=0.17.0",
    "hydra-core>=1.3.2",
    "lmdb>=1.6.2",
    "numpy>=2.2.6",
    "nvitop>=1.5.1",
    "pandas>=2.2.3",
    "peft>=0.15.2",
    "rdkit>=2025.3.2",
    "scikit-learn>=1.6.1",
    "torchmetrics>=1.7.2",
    "transformers[torch]>=4.52.4",
    "wandb>=0.20.0",
]
